coral-agent/main.py
#!/usr/bin/env python3
"""
Coral Protocol Voice Interface Agent
Real-time voice interface agent that coordinates communication between users and specialized agents.
Built on LiveKit with Coral Protocol integration.
"""

import asyncio
import os
import logging
from typing import Dict, List, Optional
from dataclasses import dataclass
import json

# LiveKit imports
from livekit.agents import AutoSubscribe, JobContext, WorkerOptions, cli, llm
from livekit.agents.voice_assistant import VoiceAssistant
from livekit import rtc
from livekit.plugins import openai

# Coral Protocol imports (MCP)
from mcp import ClientSession, StdioServerParameters
from mcp.client.stdio import stdio_client

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

@dataclass
class AgentCapabilities:
    """Define agent capabilities"""
    speech_to_text: bool = True
    text_to_speech: bool = True
    intent_analysis: bool = True
    blockchain_operations: bool = True
    payment_processing: bool = True

class CoralVoiceAgent:
    """Main Coral Protocol Voice Agent"""
    
    def __init__(self):
        self.capabilities = AgentCapabilities()
        self.active_sessions: Dict[str, Dict] = {}
        self.mcp_clients: Dict[str, ClientSession] = {}
        
    async def initialize_coral_connection(self):
        """Initialize connection to Coral Protocol"""
        try:
            # Connect to Coral Server via MCP
            coral_server_params = StdioServerParameters(
                command="coral-server",
                args=["--port", "8080"]
            )
            
            async with stdio_client(coral_server_params) as (read, write):
                async with ClientSession(read, write) as session:
                    # Initialize the session
                    await session.initialize()
                    self.mcp_clients["coral"] = session
                    logger.info("Connected to Coral Protocol server")
                    
        except Exception as e:
            logger.error(f"Failed to connect to Coral Protocol: {e}")
            # Fallback to mock mode for development
            logger.info("Running in mock mode - Coral Protocol not available")
    
    async def process_voice_input(self, audio_data: bytes, session_id: str) -> Dict:
        """Process voice input through Coral Protocol"""
        try:
            # Convert audio to text using LiveKit
            transcription = await self._transcribe_audio(audio_data)
            
            # Route to appropriate agent via Coral Protocol
            if self.mcp_clients.get("coral"):
                result = await self._route_to_coral_agent(transcription, session_id)
            else:
                # Mock response for development
                result = await self._mock_agent_response(transcription, session_id)
            
            # Generate voice response
            audio_response = await self._generate_voice_response(result["response"])
            
            return {
                "transcription": transcription,
                "intent": result["intent"],
                "response": result["response"],
                "audio_response": audio_response,
                "session_id": session_id
            }
            
        except Exception as e:
            logger.error(f"Error processing voice input: {e}")
            return {
                "error": str(e),
                "session_id": session_id
            }
    
    async def _transcribe_audio(self, audio_data: bytes) -> str:
        """Transcribe audio using LiveKit STT"""
        # This would integrate with LiveKit's STT capabilities
        # For now, return mock transcription
        return "My NFT transaction failed and I lost 0.5 ETH"
    
    async def _route_to_coral_agent(self, transcription: str, session_id: str) -> Dict:
        """Route transcription to Coral Protocol agent"""
        try:
            coral_session = self.mcp_clients["coral"]
            
            # Send message to Coral Protocol
            response = await coral_session.call_tool(
                "coral_agent",
                {
                    "message": transcription,
                    "session_id": session_id,
                    "context": "voice_interface"
                }
            )
            
            return {
                "intent": response.get("intent", "unknown"),
                "response": response.get("response", "I understand your request."),
                "actions": response.get("actions", [])
            }
            
        except Exception as e:
            logger.error(f"Error routing to Coral agent: {e}")
            return await self._mock_agent_response(transcription, session_id)
    
    async def _mock_agent_response(self, transcription: str, session_id: str) -> Dict:
        """Mock agent response for development"""
        # Simulate agent processing
        await asyncio.sleep(0.5)
        
        # Simple intent detection
        if "transaction" in transcription.lower() and "failed" in transcription.lower():
            return {
                "intent": "failed_transaction",
                "response": "I understand your frustration. I've processed a compensation NFT and initiated a refund to your wallet.",
                "actions": ["check_transaction", "mint_compensation_nft", "initiate_refund"]
            }
        elif "nft" in transcription.lower():
            return {
                "intent": "nft_inquiry",
                "response": "I can help you with NFT operations. What would you like to do?",
                "actions": ["nft_operations"]
            }
        else:
            return {
                "intent": "general_inquiry",
                "response": "I'm here to help with your Web3 needs. How can I assist you?",
                "actions": ["general_support"]
            }
    
    async def _generate_voice_response(self, text: str) -> bytes:
        """Generate voice response using LiveKit TTS"""
        # This would integrate with LiveKit's TTS capabilities
        # For now, return mock audio data
        return b"mock_audio_data"

class CoralVoiceAssistant(VoiceAssistant):
    """LiveKit Voice Assistant with Coral Protocol integration"""
    
    def __init__(self, ctx: JobContext):
        super().__init__(ctx)
        self.coral_agent = CoralVoiceAgent()
        await self.coral_agent.initialize_coral_connection()
    
    async def on_user_speech_committed(self, user_message: str):
        """Handle user speech input"""
        logger.info(f"User said: {user_message}")
        
        # Process through Coral Protocol
        result = await self.coral_agent.process_voice_input(
            user_message.encode(), 
            self.ctx.room.name
        )
        
        if "error" not in result:
            # Generate response
            response_text = result["response"]
            await self.say(response_text)
            
            # Log the interaction
            logger.info(f"Agent response: {response_text}")
        else:
            await self.say("I'm sorry, I encountered an error processing your request.")
    
    async def on_agent_speech_committed(self, agent_message: str):
        """Handle agent speech output"""
        logger.info(f"Agent said: {agent_message}")

async def entrypoint(ctx: JobContext):
    """Main entrypoint for the Coral Voice Agent"""
    logger.info("Starting Coral Protocol Voice Agent")
    
    # Initialize the voice assistant
    assistant = CoralVoiceAssistant(ctx)
    
    # Start the assistant
    await assistant.start()

if __name__ == "__main__":
    # Configure LiveKit worker
    cli.run_app(
        WorkerOptions(
            entrypoint_fnc=entrypoint,
            auto_subscribe=AutoSubscribe.AUDIO_ONLY,
        )
    )

