coral-agent/requirements.txt
# Coral Protocol Voice Interface Agent Dependencies

# LiveKit Agents Framework
livekit-agents>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-elevenlabs>=0.8.0

# Coral Protocol / MCP
mcp>=1.0.0

# Core dependencies
asyncio-mqtt>=0.16.0
websockets>=12.0
aiohttp>=3.9.0
pydantic>=2.5.0

# Audio processing
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.11.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

